// src/types/supabase.ts
export type Database = {
  public: {
    Tables: {
      customers: {
        Row: {
          address: string | null
          created_at: string
          email: string | null
          id: number
          name: string
          notes: string | null
          phone: string | null
        }
        Insert: {
          address?: string | null
          created_at?: string
          email?: string | null
          id?: number
          name: string
          notes?: string | null
          phone?: string | null
        }
        Update: {
          address?: string | null
          created_at?: string
          email?: string | null
          id?: number
          name?: string
          notes?: string | null
          phone?: string | null
        }
        Relationships: []
      }
      inventory: {
        Row: {
          created_at: string
          id: number
          notes: string | null
          price: number
          product: number
          quantity: number
        }
        Insert: {
          created_at?: string
          id?: number
          notes?: string | null
          price: number
          product: number
          quantity: number
        }
        Update: {
          created_at?: string
          id?: number
          notes?: string | null
          price?: number
          product?: number
          quantity?: number
        }
        Relationships: [
          {
            foreignKeyName: "inventory_product_fkey"
            columns: ["product"]
            isOneToOne: false
            referencedRelation: "products"
            referencedColumns: ["id"]
          },
        ]
      }
      order_items: {
        Row: {
          created_at: string
          id: number
          order: number
          price_overwrite: number | null
          product: number
          quantity: number
        }
        Insert: {
          created_at?: string
          id?: number
          order: number
          price_overwrite?: number | null
          product: number
          quantity?: number
        }
        Update: {
          created_at?: string
          id?: number
          order?: number
          price_overwrite?: number | null
          product?: number
          quantity?: number
        }
        Relationships: [
          {
            foreignKeyName: "order_items_order_fkey"
            columns: ["order"]
            isOneToOne: false
            referencedRelation: "orders"
            referencedColumns: ["id"]
          },
          {
            foreignKeyName: "order_items_product_fkey"
            columns: ["product"]
            isOneToOne: false
            referencedRelation: "products"
            referencedColumns: ["id"]
          },
        ]
      }
      orders: {
        Row: {
          created_at: string
          customer: number
          id: number
          inventory: number
          notes: string | null
          status: Database["public"]["Enums"]["Order Status"] | null
        }
        Insert: {
          created_at?: string
          customer: number
          id?: number
          inventory: number
          notes?: string | null
          status?: Database["public"]["Enums"]["Order Status"] | null
        }
        Update: {
          created_at?: string
          customer?: number
          id?: number
          inventory?: number
          notes?: string | null
          status?: Database["public"]["Enums"]["Order Status"] | null
        }
        Relationships: [
          {
            foreignKeyName: "orders_customer_fkey"
            columns: ["customer"]
            isOneToOne: false
            referencedRelation: "customers"
            referencedColumns: ["id"]
          },
          {
            foreignKeyName: "orders_inventory_fkey"
            columns: ["inventory"]
            isOneToOne: true
            referencedRelation: "inventory"
            referencedColumns: ["id"]
          },
        ]
      }
      products: {
        Row: {
          created_at: string
          id: number
          price: number | null
          title: string
          type: Database["public"]["Enums"]["Product Type"] | null
          updated_at: string
        }
        Insert: {
          created_at?: string
          id?: number
          price?: number | null
          title: string
          type?: Database["public"]["Enums"]["Product Type"] | null
          updated_at?: string
        }
        Update: {
          created_at?: string
          id?: number
          price?: number | null
          title?: string
          type?: Database["public"]["Enums"]["Product Type"] | null
          updated_at?: string
        }
        Relationships: []
      }
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      [_ in never]: never
    }
    Enums: {
      "Order Status": "pending" | "fulfilled" | "canceled"
      "Product Type": "keys" | "tools" | "parts"
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
}

type DefaultSchema = Database[Extract<keyof Database, "public">]
// ...
export type Enums<
  DefaultSchemaEnumNameOrOptions extends
    | keyof DefaultSchema["Enums"]
    | { schema: keyof Database },
  EnumName extends DefaultSchemaEnumNameOrOptions extends {
    schema: keyof Database
  }
    ? keyof Database[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"]
    : never = never,
> = DefaultSchemaEnumNameOrOptions extends { schema: keyof Database }
  ? Database[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"][EnumName]
  : DefaultSchemaEnumNameOrOptions extends keyof DefaultSchema["Enums"]
    ? DefaultSchema["Enums"][DefaultSchemaEnumNameOrOptions]
    : never

export type CompositeTypes<
  PublicCompositeTypeNameOrOptions extends
    | keyof DefaultSchema["CompositeTypes"]
    | { schema: keyof Database },
  CompositeTypeName extends PublicCompositeTypeNameOrOptions extends {
    schema: keyof Database
  }
    ? keyof Database[PublicCompositeTypeNameOrOptions["schema"]]["CompositeTypes"]
    : never = never,
> = PublicCompositeTypeNameOrOptions extends { schema: keyof Database }
  ? Database[PublicCompositeTypeNameOrOptions["schema"]]["CompositeTypes"][CompositeTypeName]
  : PublicCompositeTypeNameOrOptions extends keyof DefaultSchema["CompositeTypes"]
    ? DefaultSchema["CompositeTypes"][PublicCompositeTypeNameOrOptions]
    : never